+++ quantforge/__init__.py
@@ -769,6 +769,7 @@ from .gk_quantile import GKQuantile
 from .remez import remez
 from .floater_hormann import floater_hormann_weights, floater_hormann_interpolate
 from .aaa import aaa
+from .prony import prony, prony_reconstruct
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1473,7 +1474,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.52.0"
+__version__ = "5.53.0"
 
 __all__ = [
     "OptionType",
@@ -2509,6 +2510,8 @@ __all__ = [
     "floater_hormann_weights",
     "floater_hormann_interpolate",
     "aaa",
+    "prony",
+    "prony_reconstruct",
     "Dual",
     "dual_derivative",
     "dual_gradient",